refactor start/end into composable
